pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discovery-agent"
version = "0.1.0"
description = "GitLab to GitHub Migration Discovery Agent - scans GitLab groups and produces inventory/readiness reports"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Discovery Agent Team"}
]
keywords = ["gitlab", "github", "migration", "discovery"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
ai = [
    "openai>=1.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "responses>=0.24.0",
]
all = [
    "openai>=1.0.0",
]

[project.scripts]
discovery-agent = "discovery_agent.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["discovery_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
